requirements.txt
claude-agent-sdk
requests
python-dotenv
anyio
pytest
